refactor(CustomedTable): use composables for internal logic
- Separated sorting, pagination, multi-selection, and actions logic into dedicated composables (`useSorting`, `usePagination`, `useSelection`, `useActions`)
- Removed `.value` in the template by leveraging Vue 3's automatic ref unwrapping
- Improved readability and maintainability of the component by organizing the code modularly
