+++ src/composables/index.ts
@@ -1,1 +1,5 @@
 export * from "./useApiRequest";
+export * from "./table/useSelection";
+export * from "./table/useSorting";
+export * from "./table/usePagination";
+export * from "./table/useActions";
