+++ src/types/TableItem.ts
@@ -1,0 +1,4 @@
+export interface TableItem {
+    _id: string;
+    [key: string]: any;
+  }
