+++ src/types/heroicons.d.ts
@@ -1,0 +1,470 @@
+declare module "@heroicons/vue/*" {
+  const content: any
+  export default content
+}
+
+declare module "@heroicons/vue/solid" {
+  export { default as AcademicCapIcon } from "@heroicons/vue/solid/esm/AcademicCapIcon.js"
+  export { default as AdjustmentsIcon } from "@heroicons/vue/solid/esm/AdjustmentsIcon.js"
+  export { default as AnnotationIcon } from "@heroicons/vue/solid/esm/AnnotationIcon.js"
+  export { default as ArchiveIcon } from "@heroicons/vue/solid/esm/ArchiveIcon.js"
+  export { default as ArrowCircleDownIcon } from "@heroicons/vue/solid/esm/ArrowCircleDownIcon.js"
+  export { default as ArrowCircleLeftIcon } from "@heroicons/vue/solid/esm/ArrowCircleLeftIcon.js"
+  export { default as ArrowCircleRightIcon } from "@heroicons/vue/solid/esm/ArrowCircleRightIcon.js"
+  export { default as ArrowCircleUpIcon } from "@heroicons/vue/solid/esm/ArrowCircleUpIcon.js"
+  export { default as ArrowDownIcon } from "@heroicons/vue/solid/esm/ArrowDownIcon.js"
+  export { default as ArrowLeftIcon } from "@heroicons/vue/solid/esm/ArrowLeftIcon.js"
+  export { default as ArrowNarrowDownIcon } from "@heroicons/vue/solid/esm/ArrowNarrowDownIcon.js"
+  export { default as ArrowNarrowLeftIcon } from "@heroicons/vue/solid/esm/ArrowNarrowLeftIcon.js"
+  export { default as ArrowNarrowRightIcon } from "@heroicons/vue/solid/esm/ArrowNarrowRightIcon.js"
+  export { default as ArrowNarrowUpIcon } from "@heroicons/vue/solid/esm/ArrowNarrowUpIcon.js"
+  export { default as ArrowRightIcon } from "@heroicons/vue/solid/esm/ArrowRightIcon.js"
+  export { default as ArrowsExpandIcon } from "@heroicons/vue/solid/esm/ArrowsExpandIcon.js"
+  export { default as ArrowSmDownIcon } from "@heroicons/vue/solid/esm/ArrowSmDownIcon.js"
+  export { default as ArrowSmLeftIcon } from "@heroicons/vue/solid/esm/ArrowSmLeftIcon.js"
+  export { default as ArrowSmRightIcon } from "@heroicons/vue/solid/esm/ArrowSmRightIcon.js"
+  export { default as ArrowSmUpIcon } from "@heroicons/vue/solid/esm/ArrowSmUpIcon.js"
+  export { default as ArrowUpIcon } from "@heroicons/vue/solid/esm/ArrowUpIcon.js"
+  export { default as AtSymbolIcon } from "@heroicons/vue/solid/esm/AtSymbolIcon.js"
+  export { default as BackspaceIcon } from "@heroicons/vue/solid/esm/BackspaceIcon.js"
+  export { default as BadgeCheckIcon } from "@heroicons/vue/solid/esm/BadgeCheckIcon.js"
+  export { default as BanIcon } from "@heroicons/vue/solid/esm/BanIcon.js"
+  export { default as BeakerIcon } from "@heroicons/vue/solid/esm/BeakerIcon.js"
+  export { default as BellIcon } from "@heroicons/vue/solid/esm/BellIcon.js"
+  export { default as BookmarkAltIcon } from "@heroicons/vue/solid/esm/BookmarkAltIcon.js"
+  export { default as BookmarkIcon } from "@heroicons/vue/solid/esm/BookmarkIcon.js"
+  export { default as BookOpenIcon } from "@heroicons/vue/solid/esm/BookOpenIcon.js"
+  export { default as BriefcaseIcon } from "@heroicons/vue/solid/esm/BriefcaseIcon.js"
+  export { default as CakeIcon } from "@heroicons/vue/solid/esm/CakeIcon.js"
+  export { default as CalculatorIcon } from "@heroicons/vue/solid/esm/CalculatorIcon.js"
+  export { default as CalendarIcon } from "@heroicons/vue/solid/esm/CalendarIcon.js"
+  export { default as CameraIcon } from "@heroicons/vue/solid/esm/CameraIcon.js"
+  export { default as CashIcon } from "@heroicons/vue/solid/esm/CashIcon.js"
+  export { default as ChartBarIcon } from "@heroicons/vue/solid/esm/ChartBarIcon.js"
+  export { default as ChartPieIcon } from "@heroicons/vue/solid/esm/ChartPieIcon.js"
+  export { default as ChartSquareBarIcon } from "@heroicons/vue/solid/esm/ChartSquareBarIcon.js"
+  export { default as ChatAlt2Icon } from "@heroicons/vue/solid/esm/ChatAlt2Icon.js"
+  export { default as ChatAltIcon } from "@heroicons/vue/solid/esm/ChatAltIcon.js"
+  export { default as ChatIcon } from "@heroicons/vue/solid/esm/ChatIcon.js"
+  export { default as CheckCircleIcon } from "@heroicons/vue/solid/esm/CheckCircleIcon.js"
+  export { default as CheckIcon } from "@heroicons/vue/solid/esm/CheckIcon.js"
+  export { default as ChevronDoubleDownIcon } from "@heroicons/vue/solid/esm/ChevronDoubleDownIcon.js"
+  export { default as ChevronDoubleLeftIcon } from "@heroicons/vue/solid/esm/ChevronDoubleLeftIcon.js"
+  export { default as ChevronDoubleRightIcon } from "@heroicons/vue/solid/esm/ChevronDoubleRightIcon.js"
+  export { default as ChevronDoubleUpIcon } from "@heroicons/vue/solid/esm/ChevronDoubleUpIcon.js"
+  export { default as ChevronDownIcon } from "@heroicons/vue/solid/esm/ChevronDownIcon.js"
+  export { default as ChevronLeftIcon } from "@heroicons/vue/solid/esm/ChevronLeftIcon.js"
+  export { default as ChevronRightIcon } from "@heroicons/vue/solid/esm/ChevronRightIcon.js"
+  export { default as ChevronUpIcon } from "@heroicons/vue/solid/esm/ChevronUpIcon.js"
+  export { default as ChipIcon } from "@heroicons/vue/solid/esm/ChipIcon.js"
+  export { default as ClipboardCheckIcon } from "@heroicons/vue/solid/esm/ClipboardCheckIcon.js"
+  export { default as ClipboardCopyIcon } from "@heroicons/vue/solid/esm/ClipboardCopyIcon.js"
+  export { default as ClipboardIcon } from "@heroicons/vue/solid/esm/ClipboardIcon.js"
+  export { default as ClipboardListIcon } from "@heroicons/vue/solid/esm/ClipboardListIcon.js"
+  export { default as ClockIcon } from "@heroicons/vue/solid/esm/ClockIcon.js"
+  export { default as CloudDownloadIcon } from "@heroicons/vue/solid/esm/CloudDownloadIcon.js"
+  export { default as CloudIcon } from "@heroicons/vue/solid/esm/CloudIcon.js"
+  export { default as CloudUploadIcon } from "@heroicons/vue/solid/esm/CloudUploadIcon.js"
+  export { default as CodeIcon } from "@heroicons/vue/solid/esm/CodeIcon.js"
+  export { default as CogIcon } from "@heroicons/vue/solid/esm/CogIcon.js"
+  export { default as CollectionIcon } from "@heroicons/vue/solid/esm/CollectionIcon.js"
+  export { default as ColorSwatchIcon } from "@heroicons/vue/solid/esm/ColorSwatchIcon.js"
+  export { default as CreditCardIcon } from "@heroicons/vue/solid/esm/CreditCardIcon.js"
+  export { default as CubeIcon } from "@heroicons/vue/solid/esm/CubeIcon.js"
+  export { default as CubeTransparentIcon } from "@heroicons/vue/solid/esm/CubeTransparentIcon.js"
+  export { default as CurrencyBangladeshiIcon } from "@heroicons/vue/solid/esm/CurrencyBangladeshiIcon.js"
+  export { default as CurrencyDollarIcon } from "@heroicons/vue/solid/esm/CurrencyDollarIcon.js"
+  export { default as CurrencyEuroIcon } from "@heroicons/vue/solid/esm/CurrencyEuroIcon.js"
+  export { default as CurrencyPoundIcon } from "@heroicons/vue/solid/esm/CurrencyPoundIcon.js"
+  export { default as CurrencyRupeeIcon } from "@heroicons/vue/solid/esm/CurrencyRupeeIcon.js"
+  export { default as CurrencyYenIcon } from "@heroicons/vue/solid/esm/CurrencyYenIcon.js"
+  export { default as CursorClickIcon } from "@heroicons/vue/solid/esm/CursorClickIcon.js"
+  export { default as DatabaseIcon } from "@heroicons/vue/solid/esm/DatabaseIcon.js"
+  export { default as DesktopComputerIcon } from "@heroicons/vue/solid/esm/DesktopComputerIcon.js"
+  export { default as DeviceMobileIcon } from "@heroicons/vue/solid/esm/DeviceMobileIcon.js"
+  export { default as DeviceTabletIcon } from "@heroicons/vue/solid/esm/DeviceTabletIcon.js"
+  export { default as DocumentAddIcon } from "@heroicons/vue/solid/esm/DocumentAddIcon.js"
+  export { default as DocumentDownloadIcon } from "@heroicons/vue/solid/esm/DocumentDownloadIcon.js"
+  export { default as DocumentDuplicateIcon } from "@heroicons/vue/solid/esm/DocumentDuplicateIcon.js"
+  export { default as DocumentIcon } from "@heroicons/vue/solid/esm/DocumentIcon.js"
+  export { default as DocumentRemoveIcon } from "@heroicons/vue/solid/esm/DocumentRemoveIcon.js"
+  export { default as DocumentReportIcon } from "@heroicons/vue/solid/esm/DocumentReportIcon.js"
+  export { default as DocumentSearchIcon } from "@heroicons/vue/solid/esm/DocumentSearchIcon.js"
+  export { default as DocumentTextIcon } from "@heroicons/vue/solid/esm/DocumentTextIcon.js"
+  export { default as DotsCircleHorizontalIcon } from "@heroicons/vue/solid/esm/DotsCircleHorizontalIcon.js"
+  export { default as DotsHorizontalIcon } from "@heroicons/vue/solid/esm/DotsHorizontalIcon.js"
+  export { default as DotsVerticalIcon } from "@heroicons/vue/solid/esm/DotsVerticalIcon.js"
+  export { default as DownloadIcon } from "@heroicons/vue/solid/esm/DownloadIcon.js"
+  export { default as DuplicateIcon } from "@heroicons/vue/solid/esm/DuplicateIcon.js"
+  export { default as EmojiHappyIcon } from "@heroicons/vue/solid/esm/EmojiHappyIcon.js"
+  export { default as EmojiSadIcon } from "@heroicons/vue/solid/esm/EmojiSadIcon.js"
+  export { default as ExclamationCircleIcon } from "@heroicons/vue/solid/esm/ExclamationCircleIcon.js"
+  export { default as ExclamationIcon } from "@heroicons/vue/solid/esm/ExclamationIcon.js"
+  export { default as ExternalLinkIcon } from "@heroicons/vue/solid/esm/ExternalLinkIcon.js"
+  export { default as EyeIcon } from "@heroicons/vue/solid/esm/EyeIcon.js"
+  export { default as EyeOffIcon } from "@heroicons/vue/solid/esm/EyeOffIcon.js"
+  export { default as FastForwardIcon } from "@heroicons/vue/solid/esm/FastForwardIcon.js"
+  export { default as FilmIcon } from "@heroicons/vue/solid/esm/FilmIcon.js"
+  export { default as FilterIcon } from "@heroicons/vue/solid/esm/FilterIcon.js"
+  export { default as FingerPrintIcon } from "@heroicons/vue/solid/esm/FingerPrintIcon.js"
+  export { default as FireIcon } from "@heroicons/vue/solid/esm/FireIcon.js"
+  export { default as FlagIcon } from "@heroicons/vue/solid/esm/FlagIcon.js"
+  export { default as FolderAddIcon } from "@heroicons/vue/solid/esm/FolderAddIcon.js"
+  export { default as FolderDownloadIcon } from "@heroicons/vue/solid/esm/FolderDownloadIcon.js"
+  export { default as FolderIcon } from "@heroicons/vue/solid/esm/FolderIcon.js"
+  export { default as FolderOpenIcon } from "@heroicons/vue/solid/esm/FolderOpenIcon.js"
+  export { default as FolderRemoveIcon } from "@heroicons/vue/solid/esm/FolderRemoveIcon.js"
+  export { default as GiftIcon } from "@heroicons/vue/solid/esm/GiftIcon.js"
+  export { default as GlobeAltIcon } from "@heroicons/vue/solid/esm/GlobeAltIcon.js"
+  export { default as GlobeIcon } from "@heroicons/vue/solid/esm/GlobeIcon.js"
+  export { default as HandIcon } from "@heroicons/vue/solid/esm/HandIcon.js"
+  export { default as HashtagIcon } from "@heroicons/vue/solid/esm/HashtagIcon.js"
+  export { default as HeartIcon } from "@heroicons/vue/solid/esm/HeartIcon.js"
+  export { default as HomeIcon } from "@heroicons/vue/solid/esm/HomeIcon.js"
+  export { default as IdentificationIcon } from "@heroicons/vue/solid/esm/IdentificationIcon.js"
+  export { default as InboxIcon } from "@heroicons/vue/solid/esm/InboxIcon.js"
+  export { default as InboxInIcon } from "@heroicons/vue/solid/esm/InboxInIcon.js"
+  export { default as InformationCircleIcon } from "@heroicons/vue/solid/esm/InformationCircleIcon.js"
+  export { default as KeyIcon } from "@heroicons/vue/solid/esm/KeyIcon.js"
+  export { default as LibraryIcon } from "@heroicons/vue/solid/esm/LibraryIcon.js"
+  export { default as LightBulbIcon } from "@heroicons/vue/solid/esm/LightBulbIcon.js"
+  export { default as LightningBoltIcon } from "@heroicons/vue/solid/esm/LightningBoltIcon.js"
+  export { default as LinkIcon } from "@heroicons/vue/solid/esm/LinkIcon.js"
+  export { default as LocationMarkerIcon } from "@heroicons/vue/solid/esm/LocationMarkerIcon.js"
+  export { default as LockClosedIcon } from "@heroicons/vue/solid/esm/LockClosedIcon.js"
+  export { default as LockOpenIcon } from "@heroicons/vue/solid/esm/LockOpenIcon.js"
+  export { default as LoginIcon } from "@heroicons/vue/solid/esm/LoginIcon.js"
+  export { default as LogoutIcon } from "@heroicons/vue/solid/esm/LogoutIcon.js"
+  export { default as MailIcon } from "@heroicons/vue/solid/esm/MailIcon.js"
+  export { default as MailOpenIcon } from "@heroicons/vue/solid/esm/MailOpenIcon.js"
+  export { default as MapIcon } from "@heroicons/vue/solid/esm/MapIcon.js"
+  export { default as MenuAlt1Icon } from "@heroicons/vue/solid/esm/MenuAlt1Icon.js"
+  export { default as MenuAlt2Icon } from "@heroicons/vue/solid/esm/MenuAlt2Icon.js"
+  export { default as MenuAlt3Icon } from "@heroicons/vue/solid/esm/MenuAlt3Icon.js"
+  export { default as MenuAlt4Icon } from "@heroicons/vue/solid/esm/MenuAlt4Icon.js"
+  export { default as MenuIcon } from "@heroicons/vue/solid/esm/MenuIcon.js"
+  export { default as MicrophoneIcon } from "@heroicons/vue/solid/esm/MicrophoneIcon.js"
+  export { default as MinusCircleIcon } from "@heroicons/vue/solid/esm/MinusCircleIcon.js"
+  export { default as MinusIcon } from "@heroicons/vue/solid/esm/MinusIcon.js"
+  export { default as MinusSmIcon } from "@heroicons/vue/solid/esm/MinusSmIcon.js"
+  export { default as MoonIcon } from "@heroicons/vue/solid/esm/MoonIcon.js"
+  export { default as MusicNoteIcon } from "@heroicons/vue/solid/esm/MusicNoteIcon.js"
+  export { default as NewspaperIcon } from "@heroicons/vue/solid/esm/NewspaperIcon.js"
+  export { default as OfficeBuildingIcon } from "@heroicons/vue/solid/esm/OfficeBuildingIcon.js"
+  export { default as PaperAirplaneIcon } from "@heroicons/vue/solid/esm/PaperAirplaneIcon.js"
+  export { default as PaperClipIcon } from "@heroicons/vue/solid/esm/PaperClipIcon.js"
+  export { default as PauseIcon } from "@heroicons/vue/solid/esm/PauseIcon.js"
+  export { default as PencilAltIcon } from "@heroicons/vue/solid/esm/PencilAltIcon.js"
+  export { default as PencilIcon } from "@heroicons/vue/solid/esm/PencilIcon.js"
+  export { default as PhoneIcon } from "@heroicons/vue/solid/esm/PhoneIcon.js"
+  export { default as PhoneIncomingIcon } from "@heroicons/vue/solid/esm/PhoneIncomingIcon.js"
+  export { default as PhoneMissedCallIcon } from "@heroicons/vue/solid/esm/PhoneMissedCallIcon.js"
+  export { default as PhoneOutgoingIcon } from "@heroicons/vue/solid/esm/PhoneOutgoingIcon.js"
+  export { default as PhotographIcon } from "@heroicons/vue/solid/esm/PhotographIcon.js"
+  export { default as PlayIcon } from "@heroicons/vue/solid/esm/PlayIcon.js"
+  export { default as PlusCircleIcon } from "@heroicons/vue/solid/esm/PlusCircleIcon.js"
+  export { default as PlusIcon } from "@heroicons/vue/solid/esm/PlusIcon.js"
+  export { default as PlusSmIcon } from "@heroicons/vue/solid/esm/PlusSmIcon.js"
+  export { default as PresentationChartBarIcon } from "@heroicons/vue/solid/esm/PresentationChartBarIcon.js"
+  export { default as PresentationChartLineIcon } from "@heroicons/vue/solid/esm/PresentationChartLineIcon.js"
+  export { default as PrinterIcon } from "@heroicons/vue/solid/esm/PrinterIcon.js"
+  export { default as PuzzleIcon } from "@heroicons/vue/solid/esm/PuzzleIcon.js"
+  export { default as QrcodeIcon } from "@heroicons/vue/solid/esm/QrcodeIcon.js"
+  export { default as QuestionMarkCircleIcon } from "@heroicons/vue/solid/esm/QuestionMarkCircleIcon.js"
+  export { default as ReceiptRefundIcon } from "@heroicons/vue/solid/esm/ReceiptRefundIcon.js"
+  export { default as ReceiptTaxIcon } from "@heroicons/vue/solid/esm/ReceiptTaxIcon.js"
+  export { default as RefreshIcon } from "@heroicons/vue/solid/esm/RefreshIcon.js"
+  export { default as ReplyIcon } from "@heroicons/vue/solid/esm/ReplyIcon.js"
+  export { default as RewindIcon } from "@heroicons/vue/solid/esm/RewindIcon.js"
+  export { default as RssIcon } from "@heroicons/vue/solid/esm/RssIcon.js"
+  export { default as SaveAsIcon } from "@heroicons/vue/solid/esm/SaveAsIcon.js"
+  export { default as SaveIcon } from "@heroicons/vue/solid/esm/SaveIcon.js"
+  export { default as ScaleIcon } from "@heroicons/vue/solid/esm/ScaleIcon.js"
+  export { default as ScissorsIcon } from "@heroicons/vue/solid/esm/ScissorsIcon.js"
+  export { default as SearchCircleIcon } from "@heroicons/vue/solid/esm/SearchCircleIcon.js"
+  export { default as SearchIcon } from "@heroicons/vue/solid/esm/SearchIcon.js"
+  export { default as SelectorIcon } from "@heroicons/vue/solid/esm/SelectorIcon.js"
+  export { default as ServerIcon } from "@heroicons/vue/solid/esm/ServerIcon.js"
+  export { default as ShareIcon } from "@heroicons/vue/solid/esm/ShareIcon.js"
+  export { default as ShieldCheckIcon } from "@heroicons/vue/solid/esm/ShieldCheckIcon.js"
+  export { default as ShieldExclamationIcon } from "@heroicons/vue/solid/esm/ShieldExclamationIcon.js"
+  export { default as ShoppingBagIcon } from "@heroicons/vue/solid/esm/ShoppingBagIcon.js"
+  export { default as ShoppingCartIcon } from "@heroicons/vue/solid/esm/ShoppingCartIcon.js"
+  export { default as SortAscendingIcon } from "@heroicons/vue/solid/esm/SortAscendingIcon.js"
+  export { default as SortDescendingIcon } from "@heroicons/vue/solid/esm/SortDescendingIcon.js"
+  export { default as SparklesIcon } from "@heroicons/vue/solid/esm/SparklesIcon.js"
+  export { default as SpeakerphoneIcon } from "@heroicons/vue/solid/esm/SpeakerphoneIcon.js"
+  export { default as StarIcon } from "@heroicons/vue/solid/esm/StarIcon.js"
+  export { default as StatusOfflineIcon } from "@heroicons/vue/solid/esm/StatusOfflineIcon.js"
+  export { default as StatusOnlineIcon } from "@heroicons/vue/solid/esm/StatusOnlineIcon.js"
+  export { default as StopIcon } from "@heroicons/vue/solid/esm/StopIcon.js"
+  export { default as SunIcon } from "@heroicons/vue/solid/esm/SunIcon.js"
+  export { default as SupportIcon } from "@heroicons/vue/solid/esm/SupportIcon.js"
+  export { default as SwitchHorizontalIcon } from "@heroicons/vue/solid/esm/SwitchHorizontalIcon.js"
+  export { default as SwitchVerticalIcon } from "@heroicons/vue/solid/esm/SwitchVerticalIcon.js"
+  export { default as TableIcon } from "@heroicons/vue/solid/esm/TableIcon.js"
+  export { default as TagIcon } from "@heroicons/vue/solid/esm/TagIcon.js"
+  export { default as TemplateIcon } from "@heroicons/vue/solid/esm/TemplateIcon.js"
+  export { default as TerminalIcon } from "@heroicons/vue/solid/esm/TerminalIcon.js"
+  export { default as ThumbDownIcon } from "@heroicons/vue/solid/esm/ThumbDownIcon.js"
+  export { default as ThumbUpIcon } from "@heroicons/vue/solid/esm/ThumbUpIcon.js"
+  export { default as TicketIcon } from "@heroicons/vue/solid/esm/TicketIcon.js"
+  export { default as TranslateIcon } from "@heroicons/vue/solid/esm/TranslateIcon.js"
+  export { default as TrashIcon } from "@heroicons/vue/solid/esm/TrashIcon.js"
+  export { default as TrendingDownIcon } from "@heroicons/vue/solid/esm/TrendingDownIcon.js"
+  export { default as TrendingUpIcon } from "@heroicons/vue/solid/esm/TrendingUpIcon.js"
+  export { default as TruckIcon } from "@heroicons/vue/solid/esm/TruckIcon.js"
+  export { default as UploadIcon } from "@heroicons/vue/solid/esm/UploadIcon.js"
+  export { default as UserAddIcon } from "@heroicons/vue/solid/esm/UserAddIcon.js"
+  export { default as UserCircleIcon } from "@heroicons/vue/solid/esm/UserCircleIcon.js"
+  export { default as UserGroupIcon } from "@heroicons/vue/solid/esm/UserGroupIcon.js"
+  export { default as UserIcon } from "@heroicons/vue/solid/esm/UserIcon.js"
+  export { default as UserRemoveIcon } from "@heroicons/vue/solid/esm/UserRemoveIcon.js"
+  export { default as UsersIcon } from "@heroicons/vue/solid/esm/UsersIcon.js"
+  export { default as VariableIcon } from "@heroicons/vue/solid/esm/VariableIcon.js"
+  export { default as VideoCameraIcon } from "@heroicons/vue/solid/esm/VideoCameraIcon.js"
+  export { default as ViewBoardsIcon } from "@heroicons/vue/solid/esm/ViewBoardsIcon.js"
+  export { default as ViewGridAddIcon } from "@heroicons/vue/solid/esm/ViewGridAddIcon.js"
+  export { default as ViewGridIcon } from "@heroicons/vue/solid/esm/ViewGridIcon.js"
+  export { default as ViewListIcon } from "@heroicons/vue/solid/esm/ViewListIcon.js"
+  export { default as VolumeOffIcon } from "@heroicons/vue/solid/esm/VolumeOffIcon.js"
+  export { default as VolumeUpIcon } from "@heroicons/vue/solid/esm/VolumeUpIcon.js"
+  export { default as WifiIcon } from "@heroicons/vue/solid/esm/WifiIcon.js"
+  export { default as XCircleIcon } from "@heroicons/vue/solid/esm/XCircleIcon.js"
+  export { default as XIcon } from "@heroicons/vue/solid/esm/XIcon.js"
+  export { default as ZoomInIcon } from "@heroicons/vue/solid/esm/ZoomInIcon.js"
+  export { default as ZoomOutIcon } from "@heroicons/vue/solid/esm/ZoomOutIcon.js"
+}
+
+declare module "@heroicons/vue/outline" {
+  export { default as AcademicCapIcon } from "@heroicons/vue/outline/esm/AcademicCapIcon.js"
+  export { default as AdjustmentsIcon } from "@heroicons/vue/outline/esm/AdjustmentsIcon.js"
+  export { default as AnnotationIcon } from "@heroicons/vue/outline/esm/AnnotationIcon.js"
+  export { default as ArchiveIcon } from "@heroicons/vue/outline/esm/ArchiveIcon.js"
+  export { default as ArrowCircleDownIcon } from "@heroicons/vue/outline/esm/ArrowCircleDownIcon.js"
+  export { default as ArrowCircleLeftIcon } from "@heroicons/vue/outline/esm/ArrowCircleLeftIcon.js"
+  export { default as ArrowCircleRightIcon } from "@heroicons/vue/outline/esm/ArrowCircleRightIcon.js"
+  export { default as ArrowCircleUpIcon } from "@heroicons/vue/outline/esm/ArrowCircleUpIcon.js"
+  export { default as ArrowDownIcon } from "@heroicons/vue/outline/esm/ArrowDownIcon.js"
+  export { default as ArrowLeftIcon } from "@heroicons/vue/outline/esm/ArrowLeftIcon.js"
+  export { default as ArrowNarrowDownIcon } from "@heroicons/vue/outline/esm/ArrowNarrowDownIcon.js"
+  export { default as ArrowNarrowLeftIcon } from "@heroicons/vue/outline/esm/ArrowNarrowLeftIcon.js"
+  export { default as ArrowNarrowRightIcon } from "@heroicons/vue/outline/esm/ArrowNarrowRightIcon.js"
+  export { default as ArrowNarrowUpIcon } from "@heroicons/vue/outline/esm/ArrowNarrowUpIcon.js"
+  export { default as ArrowRightIcon } from "@heroicons/vue/outline/esm/ArrowRightIcon.js"
+  export { default as ArrowsExpandIcon } from "@heroicons/vue/outline/esm/ArrowsExpandIcon.js"
+  export { default as ArrowSmDownIcon } from "@heroicons/vue/outline/esm/ArrowSmDownIcon.js"
+  export { default as ArrowSmLeftIcon } from "@heroicons/vue/outline/esm/ArrowSmLeftIcon.js"
+  export { default as ArrowSmRightIcon } from "@heroicons/vue/outline/esm/ArrowSmRightIcon.js"
+  export { default as ArrowSmUpIcon } from "@heroicons/vue/outline/esm/ArrowSmUpIcon.js"
+  export { default as ArrowUpIcon } from "@heroicons/vue/outline/esm/ArrowUpIcon.js"
+  export { default as AtSymbolIcon } from "@heroicons/vue/outline/esm/AtSymbolIcon.js"
+  export { default as BackspaceIcon } from "@heroicons/vue/outline/esm/BackspaceIcon.js"
+  export { default as BadgeCheckIcon } from "@heroicons/vue/outline/esm/BadgeCheckIcon.js"
+  export { default as BanIcon } from "@heroicons/vue/outline/esm/BanIcon.js"
+  export { default as BeakerIcon } from "@heroicons/vue/outline/esm/BeakerIcon.js"
+  export { default as BellIcon } from "@heroicons/vue/outline/esm/BellIcon.js"
+  export { default as BookmarkAltIcon } from "@heroicons/vue/outline/esm/BookmarkAltIcon.js"
+  export { default as BookmarkIcon } from "@heroicons/vue/outline/esm/BookmarkIcon.js"
+  export { default as BookOpenIcon } from "@heroicons/vue/outline/esm/BookOpenIcon.js"
+  export { default as BriefcaseIcon } from "@heroicons/vue/outline/esm/BriefcaseIcon.js"
+  export { default as CakeIcon } from "@heroicons/vue/outline/esm/CakeIcon.js"
+  export { default as CalculatorIcon } from "@heroicons/vue/outline/esm/CalculatorIcon.js"
+  export { default as CalendarIcon } from "@heroicons/vue/outline/esm/CalendarIcon.js"
+  export { default as CameraIcon } from "@heroicons/vue/outline/esm/CameraIcon.js"
+  export { default as CashIcon } from "@heroicons/vue/outline/esm/CashIcon.js"
+  export { default as ChartBarIcon } from "@heroicons/vue/outline/esm/ChartBarIcon.js"
+  export { default as ChartPieIcon } from "@heroicons/vue/outline/esm/ChartPieIcon.js"
+  export { default as ChartSquareBarIcon } from "@heroicons/vue/outline/esm/ChartSquareBarIcon.js"
+  export { default as ChatAlt2Icon } from "@heroicons/vue/outline/esm/ChatAlt2Icon.js"
+  export { default as ChatAltIcon } from "@heroicons/vue/outline/esm/ChatAltIcon.js"
+  export { default as ChatIcon } from "@heroicons/vue/outline/esm/ChatIcon.js"
+  export { default as CheckCircleIcon } from "@heroicons/vue/outline/esm/CheckCircleIcon.js"
+  export { default as CheckIcon } from "@heroicons/vue/outline/esm/CheckIcon.js"
+  export { default as ChevronDoubleDownIcon } from "@heroicons/vue/outline/esm/ChevronDoubleDownIcon.js"
+  export { default as ChevronDoubleLeftIcon } from "@heroicons/vue/outline/esm/ChevronDoubleLeftIcon.js"
+  export { default as ChevronDoubleRightIcon } from "@heroicons/vue/outline/esm/ChevronDoubleRightIcon.js"
+  export { default as ChevronDoubleUpIcon } from "@heroicons/vue/outline/esm/ChevronDoubleUpIcon.js"
+  export { default as ChevronDownIcon } from "@heroicons/vue/outline/esm/ChevronDownIcon.js"
+  export { default as ChevronLeftIcon } from "@heroicons/vue/outline/esm/ChevronLeftIcon.js"
+  export { default as ChevronRightIcon } from "@heroicons/vue/outline/esm/ChevronRightIcon.js"
+  export { default as ChevronUpIcon } from "@heroicons/vue/outline/esm/ChevronUpIcon.js"
+  export { default as ChipIcon } from "@heroicons/vue/outline/esm/ChipIcon.js"
+  export { default as ClipboardCheckIcon } from "@heroicons/vue/outline/esm/ClipboardCheckIcon.js"
+  export { default as ClipboardCopyIcon } from "@heroicons/vue/outline/esm/ClipboardCopyIcon.js"
+  export { default as ClipboardIcon } from "@heroicons/vue/outline/esm/ClipboardIcon.js"
+  export { default as ClipboardListIcon } from "@heroicons/vue/outline/esm/ClipboardListIcon.js"
+  export { default as ClockIcon } from "@heroicons/vue/outline/esm/ClockIcon.js"
+  export { default as CloudDownloadIcon } from "@heroicons/vue/outline/esm/CloudDownloadIcon.js"
+  export { default as CloudIcon } from "@heroicons/vue/outline/esm/CloudIcon.js"
+  export { default as CloudUploadIcon } from "@heroicons/vue/outline/esm/CloudUploadIcon.js"
+  export { default as CodeIcon } from "@heroicons/vue/outline/esm/CodeIcon.js"
+  export { default as CogIcon } from "@heroicons/vue/outline/esm/CogIcon.js"
+  export { default as CollectionIcon } from "@heroicons/vue/outline/esm/CollectionIcon.js"
+  export { default as ColorSwatchIcon } from "@heroicons/vue/outline/esm/ColorSwatchIcon.js"
+  export { default as CreditCardIcon } from "@heroicons/vue/outline/esm/CreditCardIcon.js"
+  export { default as CubeIcon } from "@heroicons/vue/outline/esm/CubeIcon.js"
+  export { default as CubeTransparentIcon } from "@heroicons/vue/outline/esm/CubeTransparentIcon.js"
+  export { default as CurrencyBangladeshiIcon } from "@heroicons/vue/outline/esm/CurrencyBangladeshiIcon.js"
+  export { default as CurrencyDollarIcon } from "@heroicons/vue/outline/esm/CurrencyDollarIcon.js"
+  export { default as CurrencyEuroIcon } from "@heroicons/vue/outline/esm/CurrencyEuroIcon.js"
+  export { default as CurrencyPoundIcon } from "@heroicons/vue/outline/esm/CurrencyPoundIcon.js"
+  export { default as CurrencyRupeeIcon } from "@heroicons/vue/outline/esm/CurrencyRupeeIcon.js"
+  export { default as CurrencyYenIcon } from "@heroicons/vue/outline/esm/CurrencyYenIcon.js"
+  export { default as CursorClickIcon } from "@heroicons/vue/outline/esm/CursorClickIcon.js"
+  export { default as DatabaseIcon } from "@heroicons/vue/outline/esm/DatabaseIcon.js"
+  export { default as DesktopComputerIcon } from "@heroicons/vue/outline/esm/DesktopComputerIcon.js"
+  export { default as DeviceMobileIcon } from "@heroicons/vue/outline/esm/DeviceMobileIcon.js"
+  export { default as DeviceTabletIcon } from "@heroicons/vue/outline/esm/DeviceTabletIcon.js"
+  export { default as DocumentAddIcon } from "@heroicons/vue/outline/esm/DocumentAddIcon.js"
+  export { default as DocumentDownloadIcon } from "@heroicons/vue/outline/esm/DocumentDownloadIcon.js"
+  export { default as DocumentDuplicateIcon } from "@heroicons/vue/outline/esm/DocumentDuplicateIcon.js"
+  export { default as DocumentIcon } from "@heroicons/vue/outline/esm/DocumentIcon.js"
+  export { default as DocumentRemoveIcon } from "@heroicons/vue/outline/esm/DocumentRemoveIcon.js"
+  export { default as DocumentReportIcon } from "@heroicons/vue/outline/esm/DocumentReportIcon.js"
+  export { default as DocumentSearchIcon } from "@heroicons/vue/outline/esm/DocumentSearchIcon.js"
+  export { default as DocumentTextIcon } from "@heroicons/vue/outline/esm/DocumentTextIcon.js"
+  export { default as DotsCircleHorizontalIcon } from "@heroicons/vue/outline/esm/DotsCircleHorizontalIcon.js"
+  export { default as DotsHorizontalIcon } from "@heroicons/vue/outline/esm/DotsHorizontalIcon.js"
+  export { default as DotsVerticalIcon } from "@heroicons/vue/outline/esm/DotsVerticalIcon.js"
+  export { default as DownloadIcon } from "@heroicons/vue/outline/esm/DownloadIcon.js"
+  export { default as DuplicateIcon } from "@heroicons/vue/outline/esm/DuplicateIcon.js"
+  export { default as EmojiHappyIcon } from "@heroicons/vue/outline/esm/EmojiHappyIcon.js"
+  export { default as EmojiSadIcon } from "@heroicons/vue/outline/esm/EmojiSadIcon.js"
+  export { default as ExclamationCircleIcon } from "@heroicons/vue/outline/esm/ExclamationCircleIcon.js"
+  export { default as ExclamationIcon } from "@heroicons/vue/outline/esm/ExclamationIcon.js"
+  export { default as ExternalLinkIcon } from "@heroicons/vue/outline/esm/ExternalLinkIcon.js"
+  export { default as EyeIcon } from "@heroicons/vue/outline/esm/EyeIcon.js"
+  export { default as EyeOffIcon } from "@heroicons/vue/outline/esm/EyeOffIcon.js"
+  export { default as FastForwardIcon } from "@heroicons/vue/outline/esm/FastForwardIcon.js"
+  export { default as FilmIcon } from "@heroicons/vue/outline/esm/FilmIcon.js"
+  export { default as FilterIcon } from "@heroicons/vue/outline/esm/FilterIcon.js"
+  export { default as FingerPrintIcon } from "@heroicons/vue/outline/esm/FingerPrintIcon.js"
+  export { default as FireIcon } from "@heroicons/vue/outline/esm/FireIcon.js"
+  export { default as FlagIcon } from "@heroicons/vue/outline/esm/FlagIcon.js"
+  export { default as FolderAddIcon } from "@heroicons/vue/outline/esm/FolderAddIcon.js"
+  export { default as FolderDownloadIcon } from "@heroicons/vue/outline/esm/FolderDownloadIcon.js"
+  export { default as FolderIcon } from "@heroicons/vue/outline/esm/FolderIcon.js"
+  export { default as FolderOpenIcon } from "@heroicons/vue/outline/esm/FolderOpenIcon.js"
+  export { default as FolderRemoveIcon } from "@heroicons/vue/outline/esm/FolderRemoveIcon.js"
+  export { default as GiftIcon } from "@heroicons/vue/outline/esm/GiftIcon.js"
+  export { default as GlobeAltIcon } from "@heroicons/vue/outline/esm/GlobeAltIcon.js"
+  export { default as GlobeIcon } from "@heroicons/vue/outline/esm/GlobeIcon.js"
+  export { default as HandIcon } from "@heroicons/vue/outline/esm/HandIcon.js"
+  export { default as HashtagIcon } from "@heroicons/vue/outline/esm/HashtagIcon.js"
+  export { default as HeartIcon } from "@heroicons/vue/outline/esm/HeartIcon.js"
+  export { default as HomeIcon } from "@heroicons/vue/outline/esm/HomeIcon.js"
+  export { default as IdentificationIcon } from "@heroicons/vue/outline/esm/IdentificationIcon.js"
+  export { default as InboxIcon } from "@heroicons/vue/outline/esm/InboxIcon.js"
+  export { default as InboxInIcon } from "@heroicons/vue/outline/esm/InboxInIcon.js"
+  export { default as InformationCircleIcon } from "@heroicons/vue/outline/esm/InformationCircleIcon.js"
+  export { default as KeyIcon } from "@heroicons/vue/outline/esm/KeyIcon.js"
+  export { default as LibraryIcon } from "@heroicons/vue/outline/esm/LibraryIcon.js"
+  export { default as LightBulbIcon } from "@heroicons/vue/outline/esm/LightBulbIcon.js"
+  export { default as LightningBoltIcon } from "@heroicons/vue/outline/esm/LightningBoltIcon.js"
+  export { default as LinkIcon } from "@heroicons/vue/outline/esm/LinkIcon.js"
+  export { default as LocationMarkerIcon } from "@heroicons/vue/outline/esm/LocationMarkerIcon.js"
+  export { default as LockClosedIcon } from "@heroicons/vue/outline/esm/LockClosedIcon.js"
+  export { default as LockOpenIcon } from "@heroicons/vue/outline/esm/LockOpenIcon.js"
+  export { default as LoginIcon } from "@heroicons/vue/outline/esm/LoginIcon.js"
+  export { default as LogoutIcon } from "@heroicons/vue/outline/esm/LogoutIcon.js"
+  export { default as MailIcon } from "@heroicons/vue/outline/esm/MailIcon.js"
+  export { default as MailOpenIcon } from "@heroicons/vue/outline/esm/MailOpenIcon.js"
+  export { default as MapIcon } from "@heroicons/vue/outline/esm/MapIcon.js"
+  export { default as MenuAlt1Icon } from "@heroicons/vue/outline/esm/MenuAlt1Icon.js"
+  export { default as MenuAlt2Icon } from "@heroicons/vue/outline/esm/MenuAlt2Icon.js"
+  export { default as MenuAlt3Icon } from "@heroicons/vue/outline/esm/MenuAlt3Icon.js"
+  export { default as MenuAlt4Icon } from "@heroicons/vue/outline/esm/MenuAlt4Icon.js"
+  export { default as MenuIcon } from "@heroicons/vue/outline/esm/MenuIcon.js"
+  export { default as MicrophoneIcon } from "@heroicons/vue/outline/esm/MicrophoneIcon.js"
+  export { default as MinusCircleIcon } from "@heroicons/vue/outline/esm/MinusCircleIcon.js"
+  export { default as MinusIcon } from "@heroicons/vue/outline/esm/MinusIcon.js"
+  export { default as MinusSmIcon } from "@heroicons/vue/outline/esm/MinusSmIcon.js"
+  export { default as MoonIcon } from "@heroicons/vue/outline/esm/MoonIcon.js"
+  export { default as MusicNoteIcon } from "@heroicons/vue/outline/esm/MusicNoteIcon.js"
+  export { default as NewspaperIcon } from "@heroicons/vue/outline/esm/NewspaperIcon.js"
+  export { default as OfficeBuildingIcon } from "@heroicons/vue/outline/esm/OfficeBuildingIcon.js"
+  export { default as PaperAirplaneIcon } from "@heroicons/vue/outline/esm/PaperAirplaneIcon.js"
+  export { default as PaperClipIcon } from "@heroicons/vue/outline/esm/PaperClipIcon.js"
+  export { default as PauseIcon } from "@heroicons/vue/outline/esm/PauseIcon.js"
+  export { default as PencilAltIcon } from "@heroicons/vue/outline/esm/PencilAltIcon.js"
+  export { default as PencilIcon } from "@heroicons/vue/outline/esm/PencilIcon.js"
+  export { default as PhoneIcon } from "@heroicons/vue/outline/esm/PhoneIcon.js"
+  export { default as PhoneIncomingIcon } from "@heroicons/vue/outline/esm/PhoneIncomingIcon.js"
+  export { default as PhoneMissedCallIcon } from "@heroicons/vue/outline/esm/PhoneMissedCallIcon.js"
+  export { default as PhoneOutgoingIcon } from "@heroicons/vue/outline/esm/PhoneOutgoingIcon.js"
+  export { default as PhotographIcon } from "@heroicons/vue/outline/esm/PhotographIcon.js"
+  export { default as PlayIcon } from "@heroicons/vue/outline/esm/PlayIcon.js"
+  export { default as PlusCircleIcon } from "@heroicons/vue/outline/esm/PlusCircleIcon.js"
+  export { default as PlusIcon } from "@heroicons/vue/outline/esm/PlusIcon.js"
+  export { default as PlusSmIcon } from "@heroicons/vue/outline/esm/PlusSmIcon.js"
+  export { default as PresentationChartBarIcon } from "@heroicons/vue/outline/esm/PresentationChartBarIcon.js"
+  export { default as PresentationChartLineIcon } from "@heroicons/vue/outline/esm/PresentationChartLineIcon.js"
+  export { default as PrinterIcon } from "@heroicons/vue/outline/esm/PrinterIcon.js"
+  export { default as PuzzleIcon } from "@heroicons/vue/outline/esm/PuzzleIcon.js"
+  export { default as QrcodeIcon } from "@heroicons/vue/outline/esm/QrcodeIcon.js"
+  export { default as QuestionMarkCircleIcon } from "@heroicons/vue/outline/esm/QuestionMarkCircleIcon.js"
+  export { default as ReceiptRefundIcon } from "@heroicons/vue/outline/esm/ReceiptRefundIcon.js"
+  export { default as ReceiptTaxIcon } from "@heroicons/vue/outline/esm/ReceiptTaxIcon.js"
+  export { default as RefreshIcon } from "@heroicons/vue/outline/esm/RefreshIcon.js"
+  export { default as ReplyIcon } from "@heroicons/vue/outline/esm/ReplyIcon.js"
+  export { default as RewindIcon } from "@heroicons/vue/outline/esm/RewindIcon.js"
+  export { default as RssIcon } from "@heroicons/vue/outline/esm/RssIcon.js"
+  export { default as SaveAsIcon } from "@heroicons/vue/outline/esm/SaveAsIcon.js"
+  export { default as SaveIcon } from "@heroicons/vue/outline/esm/SaveIcon.js"
+  export { default as ScaleIcon } from "@heroicons/vue/outline/esm/ScaleIcon.js"
+  export { default as ScissorsIcon } from "@heroicons/vue/outline/esm/ScissorsIcon.js"
+  export { default as SearchCircleIcon } from "@heroicons/vue/outline/esm/SearchCircleIcon.js"
+  export { default as SearchIcon } from "@heroicons/vue/outline/esm/SearchIcon.js"
+  export { default as SelectorIcon } from "@heroicons/vue/outline/esm/SelectorIcon.js"
+  export { default as ServerIcon } from "@heroicons/vue/outline/esm/ServerIcon.js"
+  export { default as ShareIcon } from "@heroicons/vue/outline/esm/ShareIcon.js"
+  export { default as ShieldCheckIcon } from "@heroicons/vue/outline/esm/ShieldCheckIcon.js"
+  export { default as ShieldExclamationIcon } from "@heroicons/vue/outline/esm/ShieldExclamationIcon.js"
+  export { default as ShoppingBagIcon } from "@heroicons/vue/outline/esm/ShoppingBagIcon.js"
+  export { default as ShoppingCartIcon } from "@heroicons/vue/outline/esm/ShoppingCartIcon.js"
+  export { default as SortAscendingIcon } from "@heroicons/vue/outline/esm/SortAscendingIcon.js"
+  export { default as SortDescendingIcon } from "@heroicons/vue/outline/esm/SortDescendingIcon.js"
+  export { default as SparklesIcon } from "@heroicons/vue/outline/esm/SparklesIcon.js"
+  export { default as SpeakerphoneIcon } from "@heroicons/vue/outline/esm/SpeakerphoneIcon.js"
+  export { default as StarIcon } from "@heroicons/vue/outline/esm/StarIcon.js"
+  export { default as StatusOfflineIcon } from "@heroicons/vue/outline/esm/StatusOfflineIcon.js"
+  export { default as StatusOnlineIcon } from "@heroicons/vue/outline/esm/StatusOnlineIcon.js"
+  export { default as StopIcon } from "@heroicons/vue/outline/esm/StopIcon.js"
+  export { default as SunIcon } from "@heroicons/vue/outline/esm/SunIcon.js"
+  export { default as SupportIcon } from "@heroicons/vue/outline/esm/SupportIcon.js"
+  export { default as SwitchHorizontalIcon } from "@heroicons/vue/outline/esm/SwitchHorizontalIcon.js"
+  export { default as SwitchVerticalIcon } from "@heroicons/vue/outline/esm/SwitchVerticalIcon.js"
+  export { default as TableIcon } from "@heroicons/vue/outline/esm/TableIcon.js"
+  export { default as TagIcon } from "@heroicons/vue/outline/esm/TagIcon.js"
+  export { default as TemplateIcon } from "@heroicons/vue/outline/esm/TemplateIcon.js"
+  export { default as TerminalIcon } from "@heroicons/vue/outline/esm/TerminalIcon.js"
+  export { default as ThumbDownIcon } from "@heroicons/vue/outline/esm/ThumbDownIcon.js"
+  export { default as ThumbUpIcon } from "@heroicons/vue/outline/esm/ThumbUpIcon.js"
+  export { default as TicketIcon } from "@heroicons/vue/outline/esm/TicketIcon.js"
+  export { default as TranslateIcon } from "@heroicons/vue/outline/esm/TranslateIcon.js"
+  export { default as TrashIcon } from "@heroicons/vue/outline/esm/TrashIcon.js"
+  export { default as TrendingDownIcon } from "@heroicons/vue/outline/esm/TrendingDownIcon.js"
+  export { default as TrendingUpIcon } from "@heroicons/vue/outline/esm/TrendingUpIcon.js"
+  export { default as TruckIcon } from "@heroicons/vue/outline/esm/TruckIcon.js"
+  export { default as UploadIcon } from "@heroicons/vue/outline/esm/UploadIcon.js"
+  export { default as UserAddIcon } from "@heroicons/vue/outline/esm/UserAddIcon.js"
+  export { default as UserCircleIcon } from "@heroicons/vue/outline/esm/UserCircleIcon.js"
+  export { default as UserGroupIcon } from "@heroicons/vue/outline/esm/UserGroupIcon.js"
+  export { default as UserIcon } from "@heroicons/vue/outline/esm/UserIcon.js"
+  export { default as UserRemoveIcon } from "@heroicons/vue/outline/esm/UserRemoveIcon.js"
+  export { default as UsersIcon } from "@heroicons/vue/outline/esm/UsersIcon.js"
+  export { default as VariableIcon } from "@heroicons/vue/outline/esm/VariableIcon.js"
+  export { default as VideoCameraIcon } from "@heroicons/vue/outline/esm/VideoCameraIcon.js"
+  export { default as ViewBoardsIcon } from "@heroicons/vue/outline/esm/ViewBoardsIcon.js"
+  export { default as ViewGridAddIcon } from "@heroicons/vue/outline/esm/ViewGridAddIcon.js"
+  export { default as ViewGridIcon } from "@heroicons/vue/outline/esm/ViewGridIcon.js"
+  export { default as ViewListIcon } from "@heroicons/vue/outline/esm/ViewListIcon.js"
+  export { default as VolumeOffIcon } from "@heroicons/vue/outline/esm/VolumeOffIcon.js"
+  export { default as VolumeUpIcon } from "@heroicons/vue/outline/esm/VolumeUpIcon.js"
+  export { default as WifiIcon } from "@heroicons/vue/outline/esm/WifiIcon.js"
+  export { default as XCircleIcon } from "@heroicons/vue/outline/esm/XCircleIcon.js"
+  export { default as XIcon } from "@heroicons/vue/outline/esm/XIcon.js"
+  export { default as ZoomInIcon } from "@heroicons/vue/outline/esm/ZoomInIcon.js"
+  export { default as ZoomOutIcon } from "@heroicons/vue/outline/esm/ZoomOutIcon.js"
+}
